requirements.txt
matplotlib>=3.5.0
pandas>=1.3.0
Pillow>=8.3.0
pygame>=2.1.0
openpyxl>=3.0.0 